seperated reusable part of grid into new package
